// src/app/modules/hr/user/models/user-translation.model.ts
export class UserTranslation {
    userId: string;
    languageId: string;
    idCardPlaceOfIssue: string;
    officeName: string;
    titleName: string;
    positionName: string;
    passportPlaceOfIssue: string;
    nativeCountry: string;
    residentRegister: string;
    address: string;
    branchBank: string;
    bankName: string;
    nationality: string;
}